+++ src/backend/AlplaPortal.Infrastructure/Data/Migrations/20260625154133_AddBadgePrintEventCardNumber.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace AlplaPortal.Infrastructure.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddBadgePrintEventCardNumber : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "CardNumberUsed",
+                table: "BadgePrintEvents",
+                type: "nvarchar(max)",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "CardNumberUsed",
+                table: "BadgePrintEvents");
+        }
+    }
+}